+++ tests/dragonfly/conftest.py
@@ -1,11 +1,16 @@
+"""
+Pytest fixtures to be provided for all tests without import
+"""
+
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
 import os
-import pytest
-import redis
 import subprocess
 import time
+
+import pytest
+import redis
 
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 
@@ -15,6 +20,11 @@ DRAGONFLY_PATH = os.environ.get("DRAGONFLY_HOME", os.path.join(
 
 @pytest.fixture(scope="session")
 def tmp_dir():
+    """
+    Pytest fixture to provide the test temporary directory for the session
+    where the Dragonfly executable will be run and where all test data
+    should be stored. The directory will be cleaned up at the end of a session
+    """
     tmp = TemporaryDirectory()
     yield Path(tmp.name)
     tmp.cleanup()
@@ -22,50 +32,44 @@ def tmp_dir():
 
 @pytest.fixture(scope="session")
 def test_env(tmp_dir: Path):
+    """
+    Provide the environment the Dragonfly executable is running in as a
+    python dictionary
+    """
     env = os.environ.copy()
     env["DRAGONFLY_TMP"] = str(tmp_dir)
     return env
 
-# used to define a singular set of arguments for dragonfly test
-
-
-def dfly_args(*args):
-    return pytest.mark.parametrize("df_server", [args], indirect=True)
-
-
-# used to define multiple sets of arguments to test multiple dragonfly configurations
-def dfly_multi_test_args(*args):
-    return pytest.mark.parametrize("df_server", args, indirect=True)
-
 
 @pytest.fixture(scope="class", params=[[]])
 def df_server(request, tmp_dir: Path, test_env):
-    """ Starts a single DragonflyDB process, runs only once. """
-    print("Starting DragonflyDB [{}]".format(DRAGONFLY_PATH))
+    """ Starts a single DragonflyDB process, runs once per test class. """
+    print(f"Starting DragonflyDB [{DRAGONFLY_PATH}]")
     arguments = [arg.format(**test_env) for arg in request.param]
-    p = subprocess.Popen([DRAGONFLY_PATH, *arguments],
-                         env=test_env, cwd=str(tmp_dir))
+    dfly_proc = subprocess.Popen([DRAGONFLY_PATH, *arguments],
+                                 env=test_env, cwd=str(tmp_dir))
     time.sleep(0.1)
-    return_code = p.poll()
+    return_code = dfly_proc.poll()
     if return_code is not None:
-        pytest.exit("Failed to start DragonflyDB [{}]".format(DRAGONFLY_PATH))
-        p.terminate()
+        dfly_proc.terminate()
+        pytest.exit(f"Failed to start DragonflyDB [{DRAGONFLY_PATH}]")
 
     yield
 
-    print("Terminating DragonflyDB process [{}]".format(p.pid))
+    print(f"Terminating DragonflyDB process [{dfly_proc.pid}]")
     try:
-        p.terminate()
-        outs, errs = p.communicate(timeout=15)
+        dfly_proc.terminate()
+        outs, errs = dfly_proc.communicate(timeout=15)
     except subprocess.TimeoutExpired:
         print("Unable to terminate DragonflyDB gracefully, it was killed")
-        outs, errs = p.communicate()
+        outs, errs = dfly_proc.communicate()
         print(outs)
         print(errs)
 
 
 @pytest.fixture(scope="class")
 def connection(df_server):
+    """ Creates the Redis client to interact with the Dragonfly instance """
     pool = redis.ConnectionPool(decode_responses=True)
     client = redis.Redis(connection_pool=pool)
     return client
