+++ tests/dragonfly/requirements.txt
@@ -1,0 +1,12 @@
+async-timeout==4.0.2
+attrs==22.1.0
+Deprecated==1.2.13
+iniconfig==1.1.1
+packaging==21.3
+pluggy==1.0.0
+py==1.11.0
+pyparsing==3.0.9
+pytest==7.1.2
+redis==4.3.4
+tomli==2.0.1
+wrapt==1.14.1
